Created qna contoller for helper functions
