+++ server/routes/qna.js
@@ -1,28 +1,13 @@
 const qnaRouter = require('express').Router();
 const axios = require('axios');
 const api = require('../qnaApi.js');
-
-// qnaRouter.get('/', (req, res) => {
-//   res.end('QNA ROUTER');
-// });
+const qnaController = require('./qna_controller.js');
 
 qnaRouter.get('/getProductById', (req, res) => {
   let id = req.query.id;
 
-  var receiveProductInfoById = function receiveProductInfoById(id) {
-    return new Promise((resolve, reject) => {
-      api.getProductFromHR(id, (err, result) => {
-        if (err) {
-          reject(err);
-        } else {
-          resolve(result);
-        }
-      });
-    });
-  };
-  receiveProductInfoById(id)
+  qnaController.receiveProductInfoById(id)
     .then(data => {
-      //console.log('data', data);
       res.send(data);
     })
     .catch(err => {
@@ -33,19 +18,8 @@ qnaRouter.get('/getProductById', (req, res) => {
 
 qnaRouter.get('/getQuestionsList', (req, res) =>{
   let id = req.query.id;
-  var receiveQuestionList = function receiveQuestionList(id) {
-    return new Promise((resolve, reject)=> {
-      api.getQuestionsFromHR(id, (err, result) => {
-        if (err) {
-          reject(err);
-        } else {
-          resolve(result);
-        }
-      });
-    });
-  };
 
-  receiveQuestionList(id).
+  qnaController.receiveQuestionList(id).
     then(data => {
       res.send(data);
     })
@@ -58,22 +32,10 @@ qnaRouter.get('/getQuestionsList', (req, res) =>{
 
 qnaRouter.put('/addHelpfullness', (req, res) => {
   let id = req.body.params.id;
-  var increaseHelpfullness = function increaseHelpfullness(id) {
-    return new Promise((resolve, reject) => {
-      api.addHelpfullnessAtHR(id, (err, result) => {
-        if (err) {
-          reject(err);
-        } else {
-          resolve(result);
-        }
-      });
-    });
-  };
 
-  increaseHelpfullness(id)
+  qnaController.increaseHelpfullness(id)
     .then(result => {
       console.log('qna router 75', result);
-      res.sendStatus(200);
     })
     .catch(err => {
       console.log(err);
